requirements.txt
tavily-python
langchain
langchain-openai
beautifulsoup4
requests
base58
solana
solders


